Fini pour l'asp.net

+++ BeerBook.BDD/Models/Categorie.cs
@@ -6,7 +6,7 @@ namespace BeerBook.BDD
 {
     public class Categorie
     {
-        public short Identifiant { get; set; }
+        public int Identifiant { get; set; }
         public String Libelle { get; set; }
     }
 }
